requirements.txt
accelerate
opencv-python
spaces
pillow
numpy
timm
kornia
prettytable
typing
scikit-image
huggingface_hub
transformers
streamlit
loadimg